text-mining-service/pyproject.toml
[project]
name = "text-mining-service"
version = "0.1.0"
description = "LLM-based microservice for document mining"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]",
    "boto3",
    "PyPDF2",
    "python-docx",
    "numpy",
    "pandas",
    "openpyxl",
    "lancedb",
    "python-dotenv",
    "langchain",
    "langchain-text-splitters",
    "pyjwt",
    "aiohttp",
    "requests",
    "fastapi",
    "uvicorn",
    "python-crontab",
    "python-pptx",
    "python-multipart",
    "mangum",
    "psycopg2",
    "streamlit"
]